requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-docx==1.1.0
pdfplumber==0.10.3
beautifulsoup4==4.12.2
requests==2.31.0
openai==1.3.0
numpy==1.26.2
scikit-learn==1.3.2
sqlalchemy==2.0.23
pydantic>=2.7.0
aiofiles==23.2.1
python-dotenv==1.0.0
reportlab==4.0.7

